requirements.txt
flask
pyodbc
python-dotenv
openai
requests
sqlparse>=0.4.4
